+++ byond_tools/__init__.py
@@ -1,7 +1,7 @@
 # noqa: F401
 from key_tools import canonicalize
 
-from .version import __version__
+from .__version__ import __version__
 
 version = __version__
