+++ centcom/__init__.py
@@ -1,4 +1,4 @@
-from .version import __version__
+from .__version__ import __version__
 
 version = __version__
